# services/plan-api/openapi.v1.json
{
  "components": {
    "schemas": {
      "ContentDocument": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "locale": {
            "maxLength": 20,
            "minLength": 2,
            "title": "Locale",
            "type": "string"
          },
          "schemaVersion": {
            "const": 1,
            "title": "Schemaversion",
            "type": "integer"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ContentSection"
            },
            "maxItems": 100,
            "title": "Sections",
            "type": "array"
          },
          "summary": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "schemaVersion",
          "kind",
          "title",
          "summary",
          "locale"
        ],
        "title": "ContentDocument",
        "type": "object"
      },
...
      "ContentSection": {
        "additionalProperties": false,
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ContentField"
            },
            "maxItems": 100,
            "title": "Fields",
            "type": "array"
          },
          "id": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContentItem"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "layout": {
            "enum": [
              "fields",
              "list",
              "markdown",
              "table"
            ],
            "title": "Layout",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "layout"
        ],
        "title": "ContentSection",
        "type": "object"
      },
...
      "EntryComplete": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "targetId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Targetid",
            "type": "string"
          },
          "type": {
            "const": "entry.complete",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "targetId",
          "expectedVersion",
          "type"
        ],
        "title": "EntryComplete",
        "type": "object"
      },
      "EntryCreate": {
        "additionalProperties": false,
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/TaskEntryDraft"
          },
          "taskId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Taskid",
            "type": "string"
          },
          "type": {
            "const": "entry.create",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "taskId",
          "draft"
        ],
        "title": "EntryCreate",
        "type": "object"
      },
      "EntryDelete": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "targetId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Targetid",
            "type": "string"
          },
          "type": {
            "const": "entry.delete",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "targetId",
          "expectedVersion",
          "type"
        ],
        "title": "EntryDelete",
        "type": "object"
      },
      "EntryReopen": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "targetId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Targetid",
            "type": "string"
          },
          "type": {
            "const": "entry.reopen",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "targetId",
          "expectedVersion",
          "type"
        ],
        "title": "EntryReopen",
        "type": "object"
      },
      "EntrySkip": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "targetId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Targetid",
            "type": "string"
          },
          "type": {
            "const": "entry.skip",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "targetId",
          "expectedVersion",
          "type"
        ],
        "title": "EntrySkip",
        "type": "object"
      },
      "EntrySource": {
        "enum": [
          "manual",
          "rule_generated",
          "hermes"
        ],
        "title": "EntrySource",
        "type": "string"
      },
      "EntryStatus": {
        "enum": [
          "pending",
          "completed",
          "skipped"
        ],
        "title": "EntryStatus",
        "type": "string"
      },
      "EntryUpdate": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "patch": {
            "$ref": "#/components/schemas/EntryUpdatePatch"
          },
          "targetId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Targetid",
            "type": "string"
          },
          "type": {
            "const": "entry.update",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "targetId",
          "expectedVersion",
          "patch"
        ],
        "title": "EntryUpdate",
        "type": "object"
      },
      "EntryUpdatePatch": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentDocument"
              },
              {
                "type": "null"
              }
            ]
          },
          "generationRevision": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generationrevision"
          },
          "isOverridden": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isoverridden"
          },
          "scheduledDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduleddate"
          },
          "slotKey": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slotkey"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntrySource"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "EntryUpdatePatch",
        "type": "object"
      },
      "GenerationMode": {
        "enum": [
          "fixed",
          "rolling"
        ],
        "title": "GenerationMode",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MutationResult": {
        "additionalProperties": false,
        "properties": {
          "changedEntryIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Changedentryids",
            "type": "array"
          },
          "changedTaskIds": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Changedtaskids",
            "type": "array"
          },
          "serverRevision": {
            "minimum": 0.0,
            "title": "Serverrevision",
            "type": "integer"
          }
        },
        "required": [
          "serverRevision"
        ],
        "title": "MutationResult",
        "type": "object"
      },
      "ProposalConfirmResult": {
        "additionalProperties": false,
        "properties": {
          "mutation": {
            "$ref": "#/components/schemas/MutationResult"
          },
          "proposal": {
            "$ref": "#/components/schemas/ProposalView"
          }
        },
        "required": [
          "proposal",
          "mutation"
        ],
        "title": "ProposalConfirmResult",
        "type": "object"
      },
      "ProposalCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "externalSessionId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Externalsessionid",
            "type": "string"
          },
          "idempotencyKey": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Idempotencykey",
            "type": "string"
          },
          "operations": {
            "items": {
              "discriminator": {
                "mapping": {
                  "entry.complete": "#/components/schemas/EntryComplete",
                  "entry.create": "#/components/schemas/EntryCreate",
                  "entry.delete": "#/components/schemas/EntryDelete",
                  "entry.reopen": "#/components/schemas/EntryReopen",
                  "entry.skip": "#/components/schemas/EntrySkip",
                  "entry.update": "#/components/schemas/EntryUpdate",
                  "task.create": "#/components/schemas/TaskCreate",
                  "task.delete": "#/components/schemas/TaskDelete",
                  "task.setStatus": "#/components/schemas/TaskSetStatus",
                  "task.update": "#/components/schemas/TaskUpdate"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TaskCreate"
                },
                {
                  "$ref": "#/components/schemas/TaskUpdate"
                },
                {
                  "$ref": "#/components/schemas/TaskSetStatus"
                },
                {
                  "$ref": "#/components/schemas/TaskDelete"
                },
                {
                  "$ref": "#/components/schemas/EntryCreate"
                },
                {
                  "$ref": "#/components/schemas/EntryUpdate"
                },
                {
                  "$ref": "#/components/schemas/EntryComplete"
                },
                {
                  "$ref": "#/components/schemas/EntryReopen"
                },
                {
                  "$ref": "#/components/schemas/EntrySkip"
                },
                {
                  "$ref": "#/components/schemas/EntryDelete"
                }
              ]
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Operations",
            "type": "array"
          },
          "summary": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "targetTaskId": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targettaskid"
          },
          "targetVersion": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetversion"
          }
        },
        "required": [
          "externalSessionId",
          "idempotencyKey",
          "summary",
          "operations"
        ],
        "title": "ProposalCreateRequest",
        "type": "object"
      },
      "ProposalStatus": {
        "enum": [
          "pending",
          "applied",
          "cancelled",
          "expired",
          "failed"
        ],
        "title": "ProposalStatus",
        "type": "string"
      },
      "ProposalTransitionRequest": {
        "additionalProperties": false,
        "properties": {
          "externalSessionId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Externalsessionid",
            "type": "string"
          },
          "idempotencyKey": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Idempotencykey",
            "type": "string"
          }
        },
        "required": [
          "externalSessionId",
          "idempotencyKey"
        ],
        "title": "ProposalTransitionRequest",
        "type": "object"
      },
      "ProposalView": {
        "additionalProperties": false,
        "properties": {
          "appliedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appliedat"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "externalSessionId": {
            "title": "Externalsessionid",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operations": {
            "items": {
              "discriminator": {
                "mapping": {
                  "entry.complete": "#/components/schemas/EntryComplete",
                  "entry.create": "#/components/schemas/EntryCreate",
                  "entry.delete": "#/components/schemas/EntryDelete",
                  "entry.reopen": "#/components/schemas/EntryReopen",
                  "entry.skip": "#/components/schemas/EntrySkip",
                  "entry.update": "#/components/schemas/EntryUpdate",
                  "task.create": "#/components/schemas/TaskCreate",
                  "task.delete": "#/components/schemas/TaskDelete",
                  "task.setStatus": "#/components/schemas/TaskSetStatus",
                  "task.update": "#/components/schemas/TaskUpdate"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TaskCreate"
                },
                {
                  "$ref": "#/components/schemas/TaskUpdate"
                },
                {
                  "$ref": "#/components/schemas/TaskSetStatus"
                },
                {
                  "$ref": "#/components/schemas/TaskDelete"
                },
                {
                  "$ref": "#/components/schemas/EntryCreate"
                },
                {
                  "$ref": "#/components/schemas/EntryUpdate"
                },
                {
                  "$ref": "#/components/schemas/EntryComplete"
                },
                {
                  "$ref": "#/components/schemas/EntryReopen"
                },
                {
                  "$ref": "#/components/schemas/EntrySkip"
                },
                {
                  "$ref": "#/components/schemas/EntryDelete"
                }
              ]
            },
            "title": "Operations",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "targetTaskId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targettaskid"
          },
          "targetVersion": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetversion"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "externalSessionId",
          "summary",
          "operations",
          "status",
          "expiresAt",
          "targetTaskId",
          "targetVersion",
          "createdAt",
          "updatedAt",
          "appliedAt"
        ],
        "title": "ProposalView",
        "type": "object"
      },
      "ScheduleRuleV1": {
        "additionalProperties": false,
        "properties": {
          "horizonDays": {
            "const": 7,
            "title": "Horizondays",
            "type": "integer"
          },
          "schemaVersion": {
            "const": 1,
            "title": "Schemaversion",
            "type": "integer"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/ScheduleSlotV1"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Slots",
            "type": "array"
          },
          "timezone": {
            "const": "Asia/Shanghai",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "schemaVersion",
          "timezone",
          "horizonDays",
          "slots"
        ],
        "title": "ScheduleRuleV1",
        "type": "object"
      },
...
      "TaskCreate": {
        "additionalProperties": false,
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/TaskDraft"
          },
          "type": {
            "const": "task.create",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "draft"
        ],
        "title": "TaskCreate",
        "type": "object"
      },
      "TaskDelete": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "targetId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Targetid",
            "type": "string"
          },
          "type": {
            "const": "task.delete",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "targetId",
          "expectedVersion"
        ],
        "title": "TaskDelete",
        "type": "object"
      },
      "TaskDraft": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "$ref": "#/components/schemas/ContentDocument"
          },
          "entries": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TaskEntryDraft"
            },
            "title": "Entries",
            "type": "array"
          },
          "generated_through_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Through Date"
          },
          "generation_mode": {
            "$ref": "#/components/schemas/GenerationMode"
          },
          "kind": {
            "$ref": "#/components/schemas/TaskKind"
          },
          "rule_revision": {
            "default": 1,
            "title": "Rule Revision",
            "type": "integer"
          },
          "schedule_rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleRuleV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "active"
          }
        },
        "required": [
          "kind",
          "generation_mode",
          "content"
        ],
        "title": "TaskDraft",
        "type": "object"
      },
      "TaskEntryDraft": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "content": {
            "$ref": "#/components/schemas/ContentDocument"
          },
          "generation_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Revision"
          },
          "is_overridden": {
            "default": false,
            "title": "Is Overridden",
            "type": "boolean"
          },
          "scheduled_date": {
            "format": "date",
            "title": "Scheduled Date",
            "type": "string"
          },
          "slot_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Key"
          },
          "source": {
            "$ref": "#/components/schemas/EntrySource",
            "default": "manual"
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus",
            "default": "pending"
          }
        },
        "required": [
          "scheduled_date",
          "content"
        ],
        "title": "TaskEntryDraft",
        "type": "object"
      },
...
      "TaskKind": {
        "enum": [
          "daily",
          "cycle"
        ],
        "title": "TaskKind",
        "type": "string"
      },
      "TaskSetStatus": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "targetId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Targetid",
            "type": "string"
          },
          "type": {
            "const": "task.setStatus",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "targetId",
          "expectedVersion",
          "status"
        ],
        "title": "TaskSetStatus",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "active",
          "paused",
          "archived"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskUpdate": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "patch": {
            "$ref": "#/components/schemas/TaskUpdatePatch"
          },
          "targetId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Targetid",
            "type": "string"
          },
          "type": {
            "const": "task.update",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "targetId",
          "expectedVersion",
          "patch"
        ],
        "title": "TaskUpdate",
        "type": "object"
      },
      "TaskUpdatePatch": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentDocument"
              },
              {
                "type": "null"
              }
            ]
          },
          "generatedThroughDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedthroughdate"
          },
          "generationMode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "scheduleRule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleRuleV1"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "TaskUpdatePatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hermes Todo Plan API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/proposals": {
      "post": {
        "operationId": "create_proposal_v1_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Proposal"
      }
    },
    "/v1/proposals/{proposal_id}": {
      "get": {
        "operationId": "get_proposal_v1_proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "externalSessionId",
            "required": true,
            "schema": {
              "title": "Externalsessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Proposal"
      }
    },
    "/v1/proposals/{proposal_id}/cancel": {
      "post": {
        "operationId": "cancel_proposal_v1_proposals__proposal_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Proposal"
      }
    },
    "/v1/proposals/{proposal_id}/confirm": {
      "post": {
        "operationId": "confirm_proposal_v1_proposals__proposal_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalConfirmResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Proposal"
      }
    },
...
  }
}